// frontend/src/client/types.gen.ts
export type AdminProfileResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Email
     */
    email: string;
    /**
     * Username
     */
    username: string;
    /**
     * Is Active
     */
    is_active: boolean;
    /**
     * Is Verified
     */
    is_verified: boolean;
    /**
     * Created At
     */
    created_at: string;
};
// ...
export type PendingAdminResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Email
     */
    email: string;
    /**
     * Username
     */
    username: string;
    /**
     * Created At
     */
    created_at: string;
};